requirements.txt
requests
openai
python-dotenv
beautifulsoup4
